one more fix handled

+++ philo/includes/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: umut <umut@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/03 11:21:08 by usogukpi          #+#    #+#             */
-/*   Updated: 2025/02/04 23:08:40 by umut             ###   ########.fr       */
+/*   Updated: 2025/02/04 23:57:11 by umut             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,7 +73,7 @@ typedef struct s_philo
 {
 	int				id;
 	int				eaten_amount;
-	int				starvation;
+	double			starvation;
 	t_status		status;
 	t_data			*data;
 	t_death			*death;
